GL-385: move check_moving_requirements into GreenLane module
