+++ app/views/green_lanes/check_moving_requirements/edit.html.erb
@@ -22,7 +22,7 @@
 
     <%= form_for @check_moving_requirements_form, builder: GOVUKDesignSystemFormBuilder::FormBuilder,
           method: :put,
-          url: check_moving_requirements_path do |f| %>
+          url: green_lanes_check_moving_requirements_path do |f| %>
 
       <p>
         Commodity codes are internationally recognised reference numbers that describe specific goods.
