+++ app/views/green_lanes/check_moving_requirements/start.html.erb
@@ -47,7 +47,7 @@
       </ul>
     </p>
 
-    <%= button_to "Start now", edit_check_moving_requirements_path, 
+    <%= button_to "Start now", edit_green_lanes_check_moving_requirements_path, 
       method: :get,
       params: { code: @commodity_code },
       class: "govuk-button" %>
